pyproject.toml
[tool.poetry]
name = "dilivery-app"
version = "0.1.0"
description = "Dilivery app on fast api"
authors = ["gh0st <maratka2202@gmail.com>"]
license = "GNU 3"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.0"
aiosqlite = "^0.20.0"
pydantic-settings = "^2.4.0"
pytest = "^8.3.2"
alembic = "^1.13.2"
uvicorn = "^0.30.6"
asyncpg = "^0.29.0"
pyjwt = "^2.9.0"
bcrypt = "^4.2.0"
pydantic = {extras = ["email"], version = "^2.9.1"}
cryptography = "^43.0.1"
python-multipart = "^0.0.9"
pytest-asyncio = "^0.25.3"
httpx = "^0.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
